pyproject.toml
# pyproject.toml

[project]
name = "expense-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "google-adk",
    "pydantic",
    "pydantic-settings",
    "pyyaml",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "UP",
]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "google.adk.*"
ignore_missing_imports = true